PDD streamlit deploy/requirements.txt
tensorflow==2.3.0
streamlit==0.68.0
numpy==1.18.5
pillow==7.0.0